Version 669
Double precision variables (real*8) were added to all routines with this revision per Todd Campbell/Phil Gassman email
dated April 6, 2018 - Subject: SWAT double precision.  (code merged with:  d:\gassman_todd_double_precision_code041018)

-- readatmodep.f - reads for case 0 (annual) and case 1 (monthly) were changed to free format;

+++ autoirr.f
@@ -78,7 +78,7 @@
       use parm
 
       integer :: j, k
-      real :: vmma, vmm, cnv, vol, vmms, vmmd
+      real*8 :: vmma, vmm, cnv, vol, vmms, vmmd
 
       j = 0
       j = ihru
