+++ graze.f
@@ -171,7 +171,7 @@
       use parm
 
       integer :: j, l, it
-      real :: dmi, dmii, gc, gc1, swf, frt_t, xx
+      real*8 :: dmi, dmii, gc, gc1, swf, frt_t, xx
 
       j = 0
       j = ihru
