+++ pond.f
@@ -127,9 +127,9 @@
       use parm
 
       integer, intent (in) :: k
-      real :: vol, sed, pndsa, xx, targ, tpco, phosk, nitrok, chlaco
+      real*8 :: vol, sed, pndsa, xx, targ, tpco, phosk, nitrok, chlaco
       integer :: iseas
-	real :: san, sil, cla, sag, lag, inised, finsed,setsed,remsetsed
+	real*8 :: san, sil, cla, sag, lag, inised, finsed,setsed,remsetsed
 
 
         !! store initial values
